requirements.txt
Django==3.2.12
django-storages[google]==1.12.3
django-environ==0.8.1
psycopg2-binary==2.9.3
gunicorn==20.1.0
google-cloud-secret-manager==2.10.0
environ==1.0
Pillow==9.0.1
pillow-heif==0.2.0
python-docx==0.8.11
google-cloud-vision==2.7.2
# aspose-words==22.4.0